requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
redis>=5.0.0
aiohttp>=3.9.0